Move PartialStateConflictError to api.errors because of circular dep
