+++ synapse/replication/http/membership.py
@@ -17,11 +17,11 @@ from typing import TYPE_CHECKING, List, Optional, Tuple
 
 from twisted.web.server import Request
 
+from synapse.api.errors import PartialStateConflictError
 from synapse.http.server import HttpServer
 from synapse.http.servlet import parse_json_object_from_request
 from synapse.http.site import SynapseRequest
 from synapse.replication.http._base import ReplicationEndpoint
-from synapse.storage.databases.main.events import PartialStateConflictError
 from synapse.types import JsonDict, Requester, UserID
 from synapse.util.distributor import user_left_room
